feat: remove flags select
