+++ frontend/src/components/Property/Modals/EditPropertyModal.tsx
@@ -14,10 +14,9 @@ import { IconHome, IconPencil } from '@tabler/icons';
 import { useForm } from 'react-hook-form';
 import { Property, PropertyUpdateDto } from '../../../types/property';
 import GenericInput from '../../Form/GenericInput';
-import { useEffect, useState } from 'react';
-import countryToCurrency from 'country-to-currency';
+import { useEffect } from 'react';
 import getSymbolFromCurrency from 'currency-symbol-map';
-import CountrySelect from 'react-flags-select';
+import CurrencySelect from '../../Form/CurrencySelect';
 
 interface EditPropertyModalProps {
     isOpen: boolean;
@@ -59,16 +58,6 @@ function EditPropertyModal({
     }, [property]);
 
     const currency = watch('currency');
-    const [country, setCountry] = useState('US');
-
-    useEffect(() => {
-        if (country && Object.keys(countryToCurrency).includes(country)) {
-            setValue(
-                'currency',
-                countryToCurrency[country as keyof typeof countryToCurrency],
-            );
-        }
-    }, [country]);
 
     return (
         <Modal isOpen={isOpen} onClose={onClose}>
@@ -117,10 +106,12 @@ function EditPropertyModal({
                                 }}
                                 rightAdornment={getSymbolFromCurrency(currency)}
                             />
-                            <CountrySelect
-                                searchable
-                                selected={country}
-                                onSelect={(country) => setCountry(country)}
+                            <CurrencySelect
+                                name="currency"
+                                value={currency}
+                                onChange={(value) =>
+                                    setValue('currency', value)
+                                }
                             />
                             <Box w="100%">
                                 <Button
